refactor(agents): rename citation_faithfulness.py → citation_faithfulness_agent.py
Aligns with the _agent suffix convention used by all other agent modules.
The class is already named FaithfulnessAuditAgent — the filename was the
only outlier.

- Rename source and test file
- Update 19 import paths across cli/audit.py, faithfulness_cache.py,
  orchestrator.py, http_server.py, and all test files
- No behaviour change
